Properly complete dummy network client
The dealer now uses a dummy player instead of a client player, which
is better for testing the network. It should be extended later in
order to allow other types of players. For instance, #set_client_id
could be overridden by inheritance to create the ClientPlayer
once the ID becomes known. More details in #154.